move internal classes to internal package
