+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/Dsl.kt
@@ -2,6 +2,12 @@ package com.freeletics.flowredux.dsl
 
 import com.freeletics.flowredux.FlowReduxLogger
 import com.freeletics.flowredux.SideEffect
+import com.freeletics.flowredux.dsl.internal.Action
+import com.freeletics.flowredux.dsl.internal.ExternalWrappedAction
+import com.freeletics.flowredux.dsl.internal.InitialStateAction
+import com.freeletics.flowredux.dsl.internal.StoreWideCollectBuilderBlock
+import com.freeletics.flowredux.dsl.internal.StoreWideCollectorBlock
+import com.freeletics.flowredux.dsl.internal.reducer
 import com.freeletics.flowredux.reduxStore
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
