+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/InStateBuilderBlock.kt
@@ -1,6 +1,13 @@
 package com.freeletics.flowredux.dsl
 
 import com.freeletics.flowredux.SideEffect
+import com.freeletics.flowredux.dsl.internal.Action
+import com.freeletics.flowredux.dsl.internal.InStateOnEnterBlock
+import com.freeletics.flowredux.dsl.internal.InStateSideEffectBuilder
+import com.freeletics.flowredux.dsl.internal.OnActionBlock
+import com.freeletics.flowredux.dsl.internal.OnActionInStateSideEffectBuilder
+import com.freeletics.flowredux.dsl.internal.OnEnterInStateSideEffectBuilder
+import com.freeletics.flowredux.dsl.internal.Working_CollectInStateBuilder
 import kotlinx.coroutines.flow.Flow
 
 // TODO @DslMarker
