+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/internal/InStateSideEffectBuilder.kt
@@ -1,4 +1,4 @@
-package com.freeletics.flowredux.dsl
+package com.freeletics.flowredux.dsl.internal
 
 import com.freeletics.flowredux.GetState
 import com.freeletics.flowredux.SideEffect
