+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/internal/OnActionInStateSideEffectBuilder.kt
@@ -1,9 +1,11 @@
 @file:Suppress("UNCHECKED_CAST")
 
-package com.freeletics.flowredux.dsl
+package com.freeletics.flowredux.dsl.internal
 
 import com.freeletics.flowredux.SideEffect
 import com.freeletics.flowredux.GetState
+import com.freeletics.flowredux.dsl.ChangeState
+import com.freeletics.flowredux.dsl.FlatMapPolicy
 import com.freeletics.flowredux.dsl.flow.flatMapWithPolicy
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.filter
