+++ dsl/src/commonMain/kotlin/com/freeletics/flowredux/dsl/internal/Working_CollectInStateSideEffectBuilder.kt
@@ -1,7 +1,9 @@
-package com.freeletics.flowredux.dsl
+package com.freeletics.flowredux.dsl.internal
 
 import com.freeletics.flowredux.SideEffect
 import com.freeletics.flowredux.GetState
+import com.freeletics.flowredux.dsl.FlatMapPolicy
+import com.freeletics.flowredux.dsl.InStateObserverBlock
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
